修改bug

+++ launcher-domain/src/main/java/com/pactera/vo/LaunThemeVo.java
@@ -13,7 +13,7 @@ public class LaunThemeVo implements Serializable {
 	private static final long serialVersionUID = 8860802962131092316L;
 
 	private @ApiModelProperty("主键id") String id;
-	private @ApiModelProperty("排序权重") Long sort;
+	private @ApiModelProperty("排序权重") Integer sort;
 	private @ApiModelProperty("最低试用版本") String version;
 	private @ApiModelProperty("分辨率长") Long longResolution;
 	private @ApiModelProperty("分辨率宽") Long wideResolution;
